Make Response::new() an inherent method.
There is no real reason to put this in an extension trait.
